+++ zcash_client_sqlite/src/wallet.rs
@@ -136,6 +136,7 @@ pub(crate) fn pool_code(pool_type: PoolType) -> i64 {
     match pool_type {
         PoolType::Transparent => 0i64,
         PoolType::Shielded(ShieldedProtocol::Sapling) => 2i64,
+        PoolType::Shielded(ShieldedProtocol::Orchard) => 3i64,
     }
 }
 
@@ -773,6 +774,11 @@ pub(crate) fn get_received_memo(
             )
             .optional()?
             .flatten(),
+        _ => {
+            return Err(SqliteClientError::UnsupportedPoolType(PoolType::Shielded(
+                note_id.protocol(),
+            )))
+        }
     };
 
     memo_bytes